Create copying_lammps-m3gnet_renaming_POSCARs_extartcing_order_of_atoms.sh
This file copies lammps-m3gnet-fully-relax lammps input script to each directory and sub-directory. It checks if POSCAR exists in each sub-directory, if it does not then it moves the POSCAR_* (generated from pymatgen) to POSCAR and extarct the order of atoms to be input into lammps-input script for running lammps. It also converts POSCAR to in.lmp for running lammps.
Key points: Make sure that lammps-m3gnet* script lies in the directory of running this script.
Folder where this srcript is being run, make sure it only has directory which are of use meaning structures only. It should not have nay other directory.